feat: Add endpoint to update user type
The code changes add a new endpoint to update the user type (role) of a specific user. The endpoint is a PUT request to `/users/{userId}` and accepts a JSON payload with the `userType` property. The user type is validated against a list of valid user types (`moderator`, `admin`, `public`). If the user type is invalid, a 400 Bad Request response is returned. If the user is not found, a 404 Not Found response is returned. Otherwise, the user type is updated in the database and a 200 OK response is returned.

This commit message follows the established conventions in the repository.